requirements.txt
pandas~=2.3.0
numpy~=2.1.3
networkx~=3.5
matplotlib~=3.10.3
scikit-learn~=1.7.0
xgboost
beautifulsoup4~=4.13.4
requests~=2.32.4
plotly~=6.2.0
dash~=3.1.1
spacy
vaderSentiment~=3.3.2
aiohttp~=3.12.13
asyncpraw~=7.8.1
feedparser~=6.0.11
python-dotenv~=1.1.1
duckdb~=1.3.1
websockets~=15.0.1
twikit~=2.3.3
redis~=6.2.0
streamlit~=1.47.0
altair~=5.5.0
seaborn~=0.13.2
nest-asyncio~=1.6.0
pyarrow~=21.0.0
tensorflow~=2.19.0
streamlit_autorefresh
joblib~=1.5.1